refactor: убрал лишнюю вложенность
