+++ src/shared/ui/UiSvg.vue
